ATR-762: moved PX1028 analysis into the new DAC declaration analyzer and updated related code
