+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/DeclarationAnalysisDac/DacAndDacExtensionDeclarationAnalyzer.cs
@@ -1,5 +1,4 @@
-﻿
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Linq;
 
 using Acuminator.Analyzers.StaticAnalysis.Dac;
@@ -9,26 +8,47 @@ using Acuminator.Utilities.Roslyn.Semantic;
 using Acuminator.Utilities.Roslyn.Semantic.Dac;
 
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Acuminator.Analyzers.StaticAnalysis.DeclarationAnalysisDac
 {
-    public class DacAndDacExtensionDeclarationAnalyzer : DacAggregatedAnalyzerBase
+	public class DacAndDacExtensionDeclarationAnalyzer : DacAggregatedAnalyzerBase
 	{
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => 
+		public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; } =
 			ImmutableArray.Create
 			(
 				Descriptors.PX1009_InheritanceFromDacExtension,
-				Descriptors.PX1011_NotSealedDacExtension
+				Descriptors.PX1011_NotSealedDacExtension,
+				Descriptors.PX1028_ConstructorInDacDeclaration
 			);
 
 		public override void Analyze(SymbolAnalysisContext context, PXContext pxContext, DacSemanticModel dacOrDacExtension)
 		{
 			context.CancellationToken.ThrowIfCancellationRequested();
+			CheckForConstructors(context, pxContext, dacOrDacExtension);
+
+			context.CancellationToken.ThrowIfCancellationRequested();
 
 			if (dacOrDacExtension.DacType == DacType.DacExtension)
 			{
 				CheckIfDacExtensionForInheritanceIssues(context, pxContext, dacOrDacExtension);
+			}
+		}
+
+		protected virtual void CheckForConstructors(SymbolAnalysisContext context, PXContext pxContext, DacSemanticModel dacOrDacExtension)
+		{
+			var dacConstructors = dacOrDacExtension.GetMemberNodes<ConstructorDeclarationSyntax>();
+
+			foreach (var constructor in dacConstructors)
+			{
+				context.CancellationToken.ThrowIfCancellationRequested();
+				var location = constructor.Identifier.GetLocation().NullIfLocationKindIsNone() ??
+							   constructor.GetLocation();
+
+				context.ReportDiagnosticWithSuppressionCheck(
+					Diagnostic.Create(Descriptors.PX1028_ConstructorInDacDeclaration, location),
+					pxContext.CodeAnalysisSettings);
 			}
 		}
